Fix in 'set_cron_schedule' and 'get_cron_schedules' functions
Cron schedules can be considered as the same if they are have not only the same pattern but the same command. There are many cron schedules can have the same pattern.

Fixed validation in 'get_cron_schedules' function by adding cron patterns to check against into function. It was accidentally forgot to add.